move generic functions from ComparisonSeries to utils
