+++ frontend/src/utils/entities/candidate_fr_2022.ts
