+++ frontend/src/utils/entity.ts
@@ -1,5 +1,5 @@
 import { RelatedEntityObject } from './types';
-import { Video } from 'src/services/openapi';
+import { Entity, Video } from 'src/services/openapi';
 
 export const videoFromRelatedEntity = (entity: RelatedEntityObject): Video => {
   return {
@@ -15,4 +15,44 @@ export const videoFromRelatedEntity = (entity: RelatedEntityObject): Video => {
     duration: entity.metadata.duration,
     video_id: entity.metadata.video_id,
   };
+};
+
+/**
+ * Return a random entity with uid not included in the `exclude` array.
+ */
+export const selectRandomEntity = (
+  entities: Array<Entity>,
+  exclude: string[]
+): Entity => {
+  const filtered = entities.filter((entity) => !exclude.includes(entity.uid));
+
+  return filtered[Math.floor(Math.random() * filtered.length)];
+};
+
+/**
+ * Return an array of uids already compared with `uid`.
+ *
+ * @param uid An entity uid.
+ * @param comparisons An array of comparisons, represented in the form 'uidA/uidB'.
+ */
+export const alreadyComparedWith = (
+  uid: string,
+  comparisons: string[]
+): string[] => {
+  const alreadyCompared: Array<string> = [];
+
+  comparisons.forEach((comparison) => {
+    const split = comparison.split('/');
+    const index = split.indexOf(uid);
+
+    if (index != -1) {
+      if (index == 0) {
+        alreadyCompared.push(split[1]);
+      } else {
+        alreadyCompared.push(split[0]);
+      }
+    }
+  });
+
+  return alreadyCompared;
 };
